Remove nexted function
